backend/requirements.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Google Gemini AI
google-generativeai==0.3.2

# Vector Store
chromadb==0.4.22
sentence-transformers==2.3.1

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Data Processing
pandas==2.2.0
numpy==1.26.3

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
tenacity==8.2.3
python-json-logger==2.0.7

# Caching
cachetools==5.3.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Code Quality
black==24.1.1
ruff==0.1.14

# 🌟 WOW FEATURES 🌟

# Visual Generation
Pillow>=10.3.0,<11.0.0
matplotlib==3.8.2
seaborn==0.13.1

# PDF Generation
reportlab==4.0.8
pypdf==3.17.4

# Charts & Visualizations
plotly==5.18.0

# Image Processing
opencv-python-headless==4.9.0.80

# Data Export
openpyxl==3.1.2
xlsxwriter==3.1.9
